pyproject.toml
[project]
name = "yinqing-agent"
version = "0.1.0"
description = "A generic agent orchestrator based on MCP and A2A protocols."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk[sql]>=0.3.0",
    "click>=8.1.8",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "pydantic>=2.11.4",
    "mcp[cli]>=1.5.0",
    "networkx>=3.4.2"  # 用于DAG处理，虽然参考代码手写了，但引入库更稳健，不过为了保持参考代码风格，我可以先不强依赖它，但列在这里无妨
]

[project.scripts]
yinqing = "yinqing.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yinqing"]

